browser_automation/mcp_route/requirements.txt
mcp
playwright
langchain-openai
python-dotenv
